requirements.txt
# VIPA License Bot
discord.py>=2.6.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
aiosqlite>=0.20.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Twilio SMS
twilio>=9.0.0
